+++ R/check_batch_effects.R
@@ -18,13 +18,16 @@
 #' \itemize{
 #' \item pval: Matrix with P-values between phenotype variables in rows and
 #' components in columns
-#' \item plot: Heatmap as \code{\link[ComplexHeatmap]{Heatmap-class}} object
+#' \item r2: Matrix with absolute adjusted r^2 values between phenotype
+#' variables in rows and components in columns
+#' \item plot: Plot with heatmaps as returned from the
+#' \code{\link[ggpubr]{ggarrange}} function
 #' }
 #'
-#' @importFrom ComplexHeatmap Heatmap
-#' @importFrom viridis magma
-#' @importFrom grid gpar
+#' @importFrom ggplotify as.grob
+#' @importFrom ggpubr ggarrange
 #' @importFrom stats lm pf
+#'
 #' @export
 #'
 #' @examples
@@ -85,29 +88,112 @@ check_batch_effects <- function(se,
     }
 
     ## remove rows with NA
-    pval = na.omit(pval)
+    ind.na = which(apply(pval, 1, function(x) {any(is.na(x))}))
+    if (length(ind.na) > 0) {
+        pval = pval[-ind.na, , drop = FALSE]
+        abs.adj.r.squared = abs.adj.r.squared[-ind.na, , drop = FALSE]
+    }
     if (nrow(pval) == 0) {
         warning("no variables with P-values!")
         return(NULL)
     }
 
-    ## heatmap
-    breaks = c(0, 10^-4, 10^-3, 10^-2, 0.05, 1)
+    ## heatmaps
+    hm.pval = plot_heatmap(matrix = pval,
+                           type = "pval",
+                           title = title)
+    hm.r2 = plot_heatmap(matrix = abs.adj.r.squared,
+                         type = "r2",
+                         title = title)
 
-    pval.cat = apply(pval, c(1, 2), function(x) {
-        cut(x, breaks = breaks)
-    })
+    ## combine
+    hm = ggarrange(as_ggplot(as.grob(hm.pval)),
+                   as_ggplot(as.grob(hm.r2)),
+                   nrow = 1, ncol = 2)
 
-#    col = c("lightyellow", rev(heat.colors(4)))
-    col = c("white", "yellow", "orange", "red", "darkred")
-    names(col) = c("(0.05,1]",
-                   "(0.01,0.05]",
-                   "(0.001,0.01]",
-                   "(0.0001,0.001]",
-                   "(0,0.0001]")
-    hm = Heatmap(pval.cat,
+    return(list(pval = pval,
+                r2 = abs.adj.r.squared,
+                plot = hm))
+}
+
+
+
+#' Plot heatmap
+#'
+#' Plot heatmap with predefined color scheme based on type.
+#'
+#' @param matrix Matrix. Values to be plotted.
+#' @param type Character. Type of heatmap ('pval' for P-values and
+#' 'r2' for absolute adjusted r^2).
+#' @param title Character. Title of the plot.
+#'
+#' @return Heatmap as \code{\link[ComplexHeatmap]{Heatmap-class}} object
+#'
+#' @importFrom circlize colorRamp2
+#' @importFrom ComplexHeatmap Heatmap
+#' @importFrom grid gpar
+#'
+#' @export
+#'
+#' @examples
+#' data("se.gene")
+#'
+#' res.pca = calculate_mds_pca(se = se.gene,
+#'                             method = "pca")
+#'
+#' col.test = c("Age.of.patient",
+#'              "Body.surface.area",
+#'              "Duration.of.psoriasis",
+#'              "Induration",
+#'              "Overall.erythema",
+#'              "Scaling",
+#'              "Sex",
+#'              "scan.date")
+#'
+#' res = check_batch_effects(se = se.gene,
+#'                           res.pca = res.pca,
+#'                           col.test = col.test)
+#'
+#' # plot P-values
+#' plot_heatmap(matrix = res$pval,
+#'              type = "pval")
+#'
+#' # plot absolute adjusted r^2
+#' plot_heatmap(matrix = res$r2,
+#'              type = "r2")
+
+plot_heatmap <- function(matrix,
+                         type,
+                         title = NULL) {
+
+    if (type == "pval") {
+        ## categorize P-values
+        breaks = c(0, 10^-4, 10^-3, 10^-2, 0.05, 1)
+        matrix = apply(matrix, c(1, 2), function(x) {
+            cut(x, breaks = breaks)
+        })
+
+        col = c("white", "yellow", "orange", "red", "darkred")
+        names(col) = c("(0.05,1]",
+                       "(0.01,0.05]",
+                       "(0.001,0.01]",
+                       "(0.0001,0.001]",
+                       "(0,0.0001]")
+
+        name = "P-value"
+
+    } else if (type == "r2") {
+        col = colorRamp2(c(0, 1),
+                         c("white", "blue"))
+        name = "|r^2|"
+
+    } else {
+        stop(paste("type", type, "not known! Use 'pval' or 'r2."))
+    }
+
+    hm = Heatmap(matrix,
                  rect_gp = gpar(col = "black"),
-                 name = "P-value",
+                 name = name,
                  col = col,
                  cluster_rows = FALSE,
                  cluster_columns = FALSE,
@@ -116,10 +202,9 @@ check_batch_effects <- function(se,
                  column_names_side = "top",
                  column_names_rot = 0,
                  column_title = title)
-    #print(hm)
 
-    return(list(pval = pval,
-                plot = hm))
+    return(hm)
+
 }